moved functions to reviews utils file
